Move testbench into studies/
